Fjern gammelt mock-oppsett og avinstaller yet-another-fetch-mock
