+++ src/mocks/api/poao-sanity.ts
@@ -1,5 +1,5 @@
 import {http, HttpResponse, RequestHandler} from 'msw';
-import {hentSystemmeldinger} from '../../mocks-old/data/systemmeldinger';
+import {hentSystemmeldinger} from '../data/systemmeldinger';
 
 export const poaoSanityHandlers: RequestHandler[] = [
     http.get('https://poao-sanity.intern.nav.no/systemmeldinger', async () => {
