+++ src/mocks/data/geografiskBosted.ts
