+++ src/mocks/data/portefolje.ts
@@ -1,9 +1,9 @@
 import {innloggetVeileder, veiledere} from './veiledere';
 import {aktiviteter, hendelserLabels} from '../../filtrering/filter-konstanter';
-import {MOCK_CONFIG, rnd} from '../utils';
 import {faker} from '@faker-js/faker/locale/nb_NO';
 import {BarnUnder18Aar, KategoriModell} from '../../model-interfaces';
 import moment from 'moment';
+import {MOCK_CONFIG, rnd} from '../utils';
 
 faker.seed(MOCK_CONFIG.seed);
