+++ src/mocks/data/portefoljestorrelser.ts
@@ -1,4 +1,5 @@
 import {innloggetVeileder, lagTilfeldigVeilederId} from './veiledere';
+
 import {rnd} from '../utils';
 
 function lagPortefoljeStorrelse() {
